+++ examples/transactional-i2c.rs
@@ -1,6 +1,4 @@
-extern crate embedded_hal;
-extern crate linux_embedded_hal;
-use embedded_hal::blocking::i2c::{Operation as I2cOperation, Transactional};
+use embedded_hal::i2c::blocking::{Operation as I2cOperation, Transactional};
 use linux_embedded_hal::I2cdev;
 
 const ADDR: u8 = 0x12;