+++ src/cdev_pin.rs
@@ -1,29 +1,33 @@
-//! Linux CDev pin type
+//! Implementation of [`embedded-hal`] digital input/output traits using a Linux CDev pin
+//!
+//! [`embedded-hal`]: https://docs.rs/embedded-hal
 
 /// Newtype around [`gpio_cdev::LineHandle`] that implements the `embedded-hal` traits
 ///
-/// [`gpio_cdev::LineHandle`]: https://docs.rs/gpio-cdev/0.2.0/gpio_cdev/struct.LineHandle.html
+/// [`gpio_cdev::LineHandle`]: https://docs.rs/gpio-cdev/0.5.0/gpio_cdev/struct.LineHandle.html
 pub struct CdevPin(pub gpio_cdev::LineHandle, gpio_cdev::LineInfo);
 
 impl CdevPin {
     /// See [`gpio_cdev::Line::request`][0] for details.
     ///
-    /// [0]: https://docs.rs/gpio-cdev/0.2.0/gpio_cdev/struct.Line.html#method.request
+    /// [0]: https://docs.rs/gpio-cdev/0.5.0/gpio_cdev/struct.Line.html#method.request
     pub fn new(handle: gpio_cdev::LineHandle) -> Result<Self, gpio_cdev::errors::Error> {
         let info = handle.line().info()?;
         Ok(CdevPin(handle, info))
     }
 
     fn get_input_flags(&self) -> gpio_cdev::LineRequestFlags {
-        let mut flags = gpio_cdev::LineRequestFlags::INPUT;
         if self.1.is_active_low() {
-            flags.insert(gpio_cdev::LineRequestFlags::ACTIVE_LOW);
+            return gpio_cdev::LineRequestFlags::INPUT | gpio_cdev::LineRequestFlags::ACTIVE_LOW;
         }
-        return flags;
+        gpio_cdev::LineRequestFlags::INPUT
     }
 
     fn get_output_flags(&self) -> gpio_cdev::LineRequestFlags {
         let mut flags = gpio_cdev::LineRequestFlags::OUTPUT;
+        if self.1.is_active_low() {
+            flags.insert(gpio_cdev::LineRequestFlags::ACTIVE_LOW);
+        }
         if self.1.is_open_drain() {
             flags.insert(gpio_cdev::LineRequestFlags::OPEN_DRAIN);
         } else if self.1.is_open_source() {
@@ -33,27 +37,50 @@ impl CdevPin {
     }
 }
 
-impl embedded_hal::digital::OutputPin for CdevPin {
-    type Error = gpio_cdev::errors::Error;
-
-    fn set_low(&mut self) -> Result<(), Self::Error> {
-        self.0.set_value(0)
-    }
-
-    fn set_high(&mut self) -> Result<(), Self::Error> {
-        self.0.set_value(1)
+/// Converts a pin state to the gpio_cdev compatible numeric value, accounting
+/// for the active_low condition.
+fn state_to_value(state: embedded_hal::digital::PinState, is_active_low: bool) -> u8 {
+    if is_active_low {
+        match state {
+            embedded_hal::digital::PinState::High => 0,
+            embedded_hal::digital::PinState::Low => 1,
+        }
+    } else {
+        match state {
+            embedded_hal::digital::PinState::High => 1,
+            embedded_hal::digital::PinState::Low => 0,
+        }
     }
 }
 
-impl embedded_hal::digital::InputPin for CdevPin {
+impl embedded_hal::digital::blocking::OutputPin for CdevPin {
+    type Error = gpio_cdev::errors::Error;
+
+    fn set_low(&mut self) -> Result<(), Self::Error> {
+        self.0.set_value(state_to_value(
+            embedded_hal::digital::PinState::Low,
+            self.1.is_active_low(),
+        ))
+    }
+
+    fn set_high(&mut self) -> Result<(), Self::Error> {
+        self.0.set_value(state_to_value(
+            embedded_hal::digital::PinState::High,
+            self.1.is_active_low(),
+        ))
+    }
+}
+
+impl embedded_hal::digital::blocking::InputPin for CdevPin {
     type Error = gpio_cdev::errors::Error;
 
     fn is_high(&self) -> Result<bool, Self::Error> {
-        if !self.1.is_active_low() {
-            self.0.get_value().map(|val| val != 0)
-        } else {
-            self.0.get_value().map(|val| val == 0)
-        }
+        self.0.get_value().map(|val| {
+            val == state_to_value(
+                embedded_hal::digital::PinState::High,
+                self.1.is_active_low(),
+            )
+        })
     }
 
     fn is_low(&self) -> Result<bool, Self::Error> {
@@ -61,7 +88,7 @@ impl embedded_hal::digital::InputPin for CdevPin {
     }
 }
 
-impl embedded_hal::digital::IoPin<CdevPin, CdevPin> for CdevPin {
+impl embedded_hal::digital::blocking::IoPin<CdevPin, CdevPin> for CdevPin {
     type Error = gpio_cdev::errors::Error;
 
     fn into_input_pin(self) -> Result<CdevPin, Self::Error> {
@@ -95,10 +122,10 @@ impl embedded_hal::digital::IoPin<CdevPin, CdevPin> for CdevPin {
 
         CdevPin::new(line.request(
             output_flags,
-            match state {
-                embedded_hal::digital::PinState::High => 1,
-                embedded_hal::digital::PinState::Low => 0,
-            },
+            state_to_value(
+                state,
+                output_flags.intersects(gpio_cdev::LineRequestFlags::ACTIVE_LOW),
+            ),
             &consumer,
         )?)
     }