+++ src/sysfs_pin.rs
@@ -1,23 +1,25 @@
-//! Linux Sysfs pin type
+//! Implementation of [`embedded-hal`] digital input/output traits using a Linux Sysfs pin
+//!
+//! [`embedded-hal`]: https://docs.rs/embedded-hal
 
 use std::path::Path;
 
 /// Newtype around [`sysfs_gpio::Pin`] that implements the `embedded-hal` traits
 ///
-/// [`sysfs_gpio::Pin`]: https://docs.rs/sysfs_gpio/0.5.1/sysfs_gpio/struct.Pin.html
+/// [`sysfs_gpio::Pin`]: https://docs.rs/sysfs_gpio/0.6.0/sysfs_gpio/struct.Pin.html
 pub struct SysfsPin(pub sysfs_gpio::Pin);
 
 impl SysfsPin {
     /// See [`sysfs_gpio::Pin::new`][0] for details.
     ///
-    /// [0]: https://docs.rs/sysfs_gpio/0.5.1/sysfs_gpio/struct.Pin.html#method.new
+    /// [0]: https://docs.rs/sysfs_gpio/0.6.0/sysfs_gpio/struct.Pin.html#method.new
     pub fn new(pin_num: u64) -> Self {
         SysfsPin(sysfs_gpio::Pin::new(pin_num))
     }
 
     /// See [`sysfs_gpio::Pin::from_path`][0] for details.
     ///
-    /// [0]: https://docs.rs/sysfs_gpio/0.5.1/sysfs_gpio/struct.Pin.html#method.from_path
+    /// [0]: https://docs.rs/sysfs_gpio/0.6.0/sysfs_gpio/struct.Pin.html#method.from_path
     pub fn from_path<P>(path: P) -> sysfs_gpio::Result<Self>
     where
         P: AsRef<Path>,
@@ -26,19 +28,27 @@ impl SysfsPin {
     }
 }
 
-impl embedded_hal::digital::OutputPin for SysfsPin {
+impl embedded_hal::digital::blocking::OutputPin for SysfsPin {
     type Error = sysfs_gpio::Error;
 
     fn set_low(&mut self) -> Result<(), Self::Error> {
-        self.0.set_value(0)
+        if self.0.get_active_low()? {
+            self.0.set_value(1)
+        } else {
+            self.0.set_value(0)
+        }
     }
 
     fn set_high(&mut self) -> Result<(), Self::Error> {
-        self.0.set_value(1)
+        if self.0.get_active_low()? {
+            self.0.set_value(0)
+        } else {
+            self.0.set_value(1)
+        }
     }
 }
 
-impl embedded_hal::digital::InputPin for SysfsPin {
+impl embedded_hal::digital::blocking::InputPin for SysfsPin {
     type Error = sysfs_gpio::Error;
 
     fn is_high(&self) -> Result<bool, Self::Error> {
@@ -54,7 +64,7 @@ impl embedded_hal::digital::InputPin for SysfsPin {
     }
 }
 
-impl embedded_hal::digital::IoPin<SysfsPin, SysfsPin> for SysfsPin {
+impl embedded_hal::digital::blocking::IoPin<SysfsPin, SysfsPin> for SysfsPin {
     type Error = sysfs_gpio::Error;
 
     fn into_input_pin(self) -> Result<SysfsPin, Self::Error> {